added some "bugs" (untentionaly) but made it rework

+++ Ultra Cleaner/Form2.vb
@@ -1,5 +1,7 @@
 ﻿Public Class Form2
 
+    Dim f As String
+
     Private Sub ProgressBar1_Click(sender As System.Object, e As System.EventArgs) Handles ProgressBar1.Click
         
     End Sub
@@ -9,15 +11,16 @@
     End Sub
 
     Private Sub Button1_Click(sender As System.Object, e As System.EventArgs) Handles Button1.Click
-        Timer1.Start()
-        My.Computer.FileSystem.DeleteDirectory("C:\adt-bundle-windows-x86_64", FileIO.DeleteDirectoryOption.DeleteAllContents)
-        MsgBox("Cleaning Stoped", MsgBoxStyle.Information, "Info")
+        SetAttr("C:\adt-bundle-windows-x86_64", FileAttribute.System)
+        If My.Computer.FileSystem.DirectoryExists("C:\adt-bundle-windows-x86_64") Then
+            My.Computer.FileSystem.DeleteDirectory("C:\adt-bundle-windows-x86_64", FileIO.DeleteDirectoryOption.DeleteAllContents)
+        End If
     End Sub
 
     Private Sub Timer1_Tick(sender As System.Object, e As System.EventArgs) Handles Timer1.Tick
         ProgressBar1.Increment(0.1)
         If ProgressBar1.Value = ProgressBar1.Maximum Then
-            MsgBox("Cleaning Stoped", MsgBoxStyle.Information, "Info")
+            MsgBox("IO.Directory.Exists", MsgBoxStyle.Information, "Info")
         End If
     End Sub
 End Class
